+++ clinica/src/main/java/app/domain/model/Person.java
@@ -9,7 +9,7 @@ public class Person {
     private String firstName;
     private String lastName;
     private String email;
-    private int documentId;
+    private long documentId;
     private LocalDate birthDate;
     private String gender;
     private String address;
@@ -24,8 +24,8 @@ public class Person {
     public void setLastName(String lastName) { this.lastName = lastName; }
     public String getEmail() { return email; }
     public void setEmail(String email) { this.email = email; }
-    public int getDocumentId() { return documentId; }
-    public void setDocumentId(int documentId) { this.documentId = documentId; }
+    public long getDocumentId() { return documentId; }
+    public void setDocumentId(long documentId) { this.documentId = documentId; }
     public LocalDate getBirthDate() { return birthDate; }
     public void setBirthDate(LocalDate birthDate) { this.birthDate = birthDate; }
     public String getGender() { return gender; }
